# docs/openapi/pdt-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Platform Data Toolkit (PDT) API",
        "description": "PDT is a centralized API server that serves as an enterprise knowledge silo for company-specific concepts, documents, and their relationships.\n\n## Features\n\n- **Knowledge Asset Management**: Store and manage knowledge assets with rich metadata\n- **Tag-Based Classification**: Multi-dimensional tagging including asset type, business domain, language, etc.\n- **Graph-Based Relationships**: Define and traverse relationships between assets\n- **Concept Collections**: Organize assets into named collections\n- **Search & Discovery**: Full-text search and tag-based filtering\n- **Audit Logging**: Track all changes with user attribution\n- **Authorization Context**: Cedar-based visibility, owner groups, and confidentiality management",
        "contact": {
            "name": "PDT Team"
        },
        "license": {
            "name": "MIT OR Apache-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/assets/{id}/graph": {
            "get": {
                "tags": [
                    "relations"
                ],
                "summary": "Traverse the relationship graph from an asset",
                "operationId": "traverse_graph",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "depth",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Graph traversal results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GraphNode"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            }
        },
...
        "/api/assets/{id}/relations": {
            "get": {
                "tags": [
                    "relations"
                ],
                "summary": "Get all relations for an asset",
                "operationId": "get_asset_relations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of relations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Relation"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            }
        },
...
        "/api/relations": {
            "post": {
                "tags": [
                    "relations"
                ],
                "summary": "Create a new relation",
                "operationId": "create_relation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRelationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Relation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Relation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Source or target asset not found"
                    }
                }
            }
        },
        "/api/relations/{id}": {
            "get": {
                "tags": [
                    "relations"
                ],
                "summary": "Get relation by ID",
                "operationId": "get_relation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Relation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relation found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Relation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "relations"
                ],
                "summary": "Delete a relation",
                "operationId": "delete_relation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Relation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relation deleted successfully"
                    },
                    "404": {
                        "description": "Relation not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateRelationRequest": {
                "type": "object",
                "description": "Request to create a new relation",
                "required": [
                    "from_asset_id",
                    "to_asset_id",
                    "relation_type"
                ],
                "properties": {
                    "from_asset_id": {
                        "type": "string"
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": {},
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "relation_type": {
                        "$ref": "#/components/schemas/RelationType"
                    },
                    "to_asset_id": {
                        "type": "string"
                    }
                }
            },
            "GraphNode": {
                "type": "object",
                "description": "Node in the relationship graph",
                "required": [
                    "asset_id",
                    "depth",
                    "relations"
                ],
                "properties": {
                    "asset_id": {
                        "type": "string"
                    },
                    "depth": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "relations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Relation"
                        }
                    }
                }
            },
...
            "Relation": {
                "type": "object",
                "description": "Relationship between two assets",
                "required": [
                    "_id",
                    "from_asset_id",
                    "to_asset_id",
                    "relation_type",
                    "created_at",
                    "created_by"
                ],
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "from_asset_id": {
                        "type": "string"
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": {},
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "relation_type": {
                        "$ref": "#/components/schemas/RelationType"
                    },
                    "to_asset_id": {
                        "type": "string"
                    }
                }
            },
            "RelationType": {
                "type": "string",
                "description": "Types of relationships between assets",
                "enum": [
                    "contains",
                    "references",
                    "related_to",
                    "depends_on",
                    "supersedes",
                    "complements"
                ]
            },
...
        }
    },
...
}